+++ CHANGELOG.md
@@ -19,6 +19,8 @@ FlatLaf Change Log
 - Show mnemonics always when a menu bar is active or a popup menu is visible.
 - Hide mnemonics if window is deactivated (e.g. <kbd>Alt+Tab</kbd> to another
   window). (issue #43)
+- macOS: Fixed NPE if using `JMenuBar` in `JInternalFrame` and macOS screen menu
+  bar is enabled (with `-Dapple.laf.useScreenMenuBar=true`). (issue #90)
 
 
 ## 0.33